+++ src/fsScanning.ts
@@ -50,23 +50,3 @@ export async function recursiveReaddir(basePath: string, callback: (path: string
         }
     }
 }
-
-export async function getNeighboringDirectories(currentPath: string) {
-    const parentPath = dirname(currentPath);
-    let fileNames = await readdirOrEmpty(parentPath);
-    assert(fileNames.length >= 1);
-
-    const neighboringDirectories = [];
-    for (const fileName of fileNames) {
-        const path = join(parentPath, fileName);
-        const stats = await statsOrUndefined(path);
-        if (stats === undefined) {
-            continue;
-        }
-        if (stats.isDirectory()) {
-            neighboringDirectories.push(path);
-        }
-    }
-
-    return neighboringDirectories;
-}